WIP: Stashing unfinished thought on a new branch
Looking for a way to reduce the amount of duplicated code checking for 'Not Found' errors and retrying as a file. This code started by trying to solve that issue with Actions. Not sure if that makes sense still but don't want to throw this code away so I'll put it on a separate branch for a while.